requirements.txt
selenium==4.10.0
pillow==9.5.0
tk==0.1.0
webdriver-manager==3.8.5